requirements.txt
python-dotenv
telethon
discord.py